+++ src/lang/lexing/index.ts
@@ -1,0 +1,9 @@
+import Lexer  from "./lexer";
+
+const runLexer = (text: string, filename: string) => {
+    const lexer = new Lexer(text, filename);
+
+    return lexer.makeTokens();
+};
+
+export default runLexer;
